--- client/src/utils/utils.ts
@@ -1,60 +1,0 @@
-import { Position, Selection, window } from "vscode";
-import { BaseLanguageClient } from "vscode-languageclient";
-
-interface FoldingBlock {
-  startLine: number;
-  startCol: number;
-  endLine: number;
-  endCol: number;
-}
-
-export async function getSelectedRegions(
-  client: BaseLanguageClient,
-): Promise<Selection[]> {
-  const result: string[] = [];
-
-  async function pushBlock(line: number, col: number) {
-    const block = await client.sendRequest<FoldingBlock>(
-      "sas/getFoldingBlock",
-      {
-        textDocument: { uri: window.activeTextEditor.document.uri.toString() },
-        line,
-        col,
-      },
-    );
-    if (block) {
-      const start = doc.offsetAt(new Position(block.startLine, block.startCol));
-      const end = doc.offsetAt(new Position(block.endLine, block.endCol));
-      const key = `${start}-${end}`;
-      if (result.indexOf(key) === -1) {
-        result.push(key);
-      }
-      return end;
-    }
-  }
-
-  const editor = window.activeTextEditor;
-  const doc = editor.document;
-  for (const selection of editor.selections) {
-    const start = doc.offsetAt(selection.start);
-    let end = doc.offsetAt(selection.end);
-    const selectedText = doc.getText(selection);
-    if (selectedText.endsWith("\n")) {
-      --end;
-    }
-    for (let i = start; i <= end; i++) {
-      const pos = doc.positionAt(i);
-      const blockEnd = await pushBlock(pos.line, pos.character);
-      if (blockEnd && blockEnd > i) {
-        i = blockEnd;
-      }
-    }
-  }
-  return result.map((key) => {
-    const [start, end] = key.split("-");
-    return new Selection(
-      doc.positionAt(parseInt(start)),
-      doc.positionAt(parseInt(end)),
-    );
-  });
-}
